fix test HTTP_405, refactor Code Signup & ChooseCampaign, handle error rgpd
